+++ R/utils.R
@@ -937,35 +937,19 @@ is_rstudio = function() rstudio_mode() != ''
 is_rstudio_server = function() rstudio_mode() == 'server'
 
 # tweak some env vars when building a site or running the hugo server
-tweak_hugo_env = function(server = TRUE) {
+tweak_hugo_env = function(baseURL = NULL, relativeURLs = NULL, server = FALSE) {
   # set baseURL properly when it doesn't contain protocol or domain:
   # https://github.com/gohugoio/hugo/issues/7823 (add example.org/ to it); or
   # when relativeURLs = true, set baseURL to /
   config = load_config()
-  b = get_config('baseurl', '/', config)
+  b = if (is.null(baseURL)) get_config('baseurl', '/', config) else baseURL
+  b = sub('^/([^/].*)', '\\1', b)
   c1 = b != '/' && !grepl('^https?://[^/]+', b)
-  c2 = get_config('relativeurls', FALSE, config)
-  if (c2 || c1) {
-    b = sub('^/', '', b)
-    if (server && c1) b = paste0('https://example.org/', b)
-    Sys.setenv(HUGO_BASEURL = if (c2) '/' else b)
-    do.call(
-      on.exit, list(substitute(Sys.unsetenv('HUGO_BASEURL')), add = TRUE),
-      envir = parent.frame()
-    )
-  }
+  c2 = if (is.null(relativeURLs)) get_config('relativeurls', FALSE, config) else relativeURLs
+  if (server && c1) b = paste0(if (grepl('^//', b)) 'http:' else 'http://example.org/', b)
 
-  if (!server) return()
-
-  # RStudio Server uses a proxy like http://localhost:8787/p/56a946ed/ for
-  # http://localhost:4321, so we must use relativeURLs = TRUE:
-  # https://github.com/rstudio/blogdown/issues/124
-  if (!is_rstudio_server()) return()
-  Sys.setenv(HUGO_RELATIVEURLS = 'true')
-  do.call(
-    on.exit, list(substitute(Sys.unsetenv('HUGO_RELATIVEURLS')), add = TRUE),
-    envir = parent.frame()
-  )
+  v = set_envvar(c(HUGO_BASEURL = if (c2) '/' else b, HUGO_RELATIVEURLS = tolower(c2)))
+  exit_call(function() set_envvar(v))
 }
 
 get_author = function() {
@@ -1061,3 +1045,24 @@ na2null = function(x, default = NULL) {
   x = x[sort(names(x))]
   x
 })
+
+# set env vars from a named character vector, and return the old values of the
+# vars, so they could be restored later
+set_envvar = function(vars) {
+  if (is.null(nms <- names(vars)) || any(nms == '')) stop(
+    "The 'vars' argument must take a named character vector."
+  )
+  vals = Sys.getenv(nms, NA, names = TRUE)
+  i = is.na(vars)
+  suppressWarnings(Sys.unsetenv(nms[i]))
+  if (length(vars <- vars[!i])) do.call(Sys.setenv, as.list(vars))
+  invisible(vals)
+}
+
+# call on.exit in a parent function
+exit_call = function(fun, n = 2) {
+  do.call(
+    on.exit, list(substitute(fun(), list(fun = fun)), add = TRUE),
+    envir = parent.frame(n)
+  )
+}